Added cpp support to webserver

+++ exchange/src/linter/runtime/cpp/cpp_runtime.hpp
@@ -7,8 +7,10 @@ namespace nutc::lint {
 class CppRuntime : public Runtime {
 public:
     CppRuntime(
-        std::string algo, LimitOrderFunction limit_order,
-        MarketOrderFunction market_order, CancelOrderFunction cancel_order
+        std::string algo,
+        LimitOrderFunction limit_order,
+        MarketOrderFunction market_order,
+        CancelOrderFunction cancel_order
     );
 
     ~CppRuntime() override;
@@ -16,27 +18,32 @@ public:
     std::optional<std::string> init() override;
 
     void fire_on_trade_update(
-        common::Ticker ticker, common::Side side, double price, double quantity
-    ) const override;
+            common::Ticker ticker, common::Side side, float price, float quantity
+            ) const override;
 
     void fire_on_orderbook_update(
-        common::Ticker ticker, common::Side side, double price, double quantity
-    ) const override;
+            common::Ticker ticker, common::Side side, float price, float quantity
+            ) const override;
 
     void fire_on_account_update(
-        common::Ticker ticker, common::Side side, double price, double quantity,
-        double capital
-    ) const override;
-
+            common::Ticker ticker,
+            common::Side side,
+            float price,
+            float quantity,
+            float capital
+            ) const override;
 private:
+
     using Strategy = void;
-    using InitFunc = Strategy* (*)(MarketOrderFunction, LimitOrderFunction,
+    using InitFunc = Strategy* (*)(MarketOrderFunction,
+                                   LimitOrderFunction,
                                    CancelOrderFunction);
     using OnTradeUpdateFunc =
-        void (*)(Strategy*, common::Ticker, common::Side, double, double);
+	void (*)(Strategy*, common::Ticker, common::Side, float, float);
     using OnOrderBookUpdateFunc = OnTradeUpdateFunc;
-    using OnAccountUpdateFunc =
-        void (*)(Strategy*, common::Ticker, common::Side, double, double, double);
+    using OnAccountUpdateFunc = void (*)(
+		Strategy*, common::Ticker, common::Side, float, float, float
+    );
 
     OnTradeUpdateFunc on_trade_update_func_;
     OnOrderBookUpdateFunc on_orderbook_update_func_;
